+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = coverage-clean,{py36,py37,pypy3}-{twlatest,twtrunk,twlowest}-{aclatest,acmaster}-alldeps,flake8,docs,coverage-report
+envlist = coverage-clean,{py36,py38,pypy3}-{twlatest,twtrunk,twlowest}-{aclatest,acmaster}-alldeps,flake8,docs,coverage-report
 
 [testenv]
 setenv =
@@ -22,7 +22,7 @@ commands =
         {envdir}/bin/trial --temp-directory={envtmpdir}/_trial_temp {posargs:txacme integration}
 
 [testenv:flake8]
-basepython = python3.6
+basepython = python3.8
 deps =
     flake8
     pep8-naming
